Rename launch root
